Attempting WebGL based Chain Reduction
The Chains are very processor intensive. A couple of days ago, a customer and I were discussing GPU implementation tools we can use in the office. In particular we delved into sort techniques (in particular parrallel bubble sorts). This got me thinking about this project and last night it occurred to me that the chain can be walked backwards using a cellular automata technique (little more than a bubble sort).