Chain algorithm uses larger sentinel values, and initianlizing values for chain processing.
Detecting the difference between 1/255 and 2/255 was proving difficult, instead the directional sentinel values were defined as 1/4. This gives a much larger space to detect which value was being used when testing for "equality".

Have also decided to use the red channel to count the lenght of the chain. This is used to determine where to start in the chain. Have to initialize the red channel to zero to do this.