Starting at highwater mark when calculating chain.
There is always a question of where the chain ends. You can spot the end of the chain by watching for the point at which the values start to continously get smaller. This indicates you are getting any matches. When working backwards this is the first point that the score stops getting bigger.