Significant error trapping.
There is a race condition that is ... interesting. If you load the program, and then clear almost right away, you end up with a timeline where you're clearing and loading overlap. Bad things happen.

I did manage to trap some of the occurrences of this, but not all.

Embarrassingly, the solution was to refresh the page and ignore the logging.